solution: Fixed demob for project 'Computer repair workshop'
